requirements.txt
weaviate-client==v4.10.0b3
